Updated some of the make-bindist scripts for x64sc.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@22533 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/sdl/make-bindist_dingoo.sh
@@ -11,11 +11,18 @@ VICEVERSION=$2
 ZIPKIND=$3
 TOPSRCDIR=$4
 
+if [ ! -e src/x64 -o ! -e src/x64dtv -o ! -e src/x64sc -o ! -e src/x128 -o ! -e src/xvic -o ! -e src/xpet -o ! -e src/xplus4 -o ! -e src/xcbm2 -o ! -e src/c1541 -o ! -e src/petcat -o ! -e src/cartconv ]
+then
+  echo Error: executable file\(s\) not found, do a \"make\" first
+  exit 1
+fi
+
 echo Generating DINGOO SDL port binary distribution.
 rm -f -r SDLVICE-dingoo-$VICEVERSION
 mkdir SDLVICE-dingoo-$VICEVERSION
 $STRIP src/x64
 $STRIP src/x64dtv
+$STRIP src/x64sc
 $STRIP src/x128
 $STRIP src/xvic
 $STRIP src/xpet
@@ -23,6 +30,7 @@ $STRIP src/xplus4
 $STRIP src/xcbm2
 cp src/x64 src/x64dtv src/x128 src/xvic SDLVICE-dingoo-$VICEVERSION
 cp src/xpet src/xplus4 src/xcbm2 SDLVICE-dingoo-$VICEVERSION
+cp src/x64sc SDLVICE-dingoo-$VICEVERSION
 cp $TOPSRCDIR/src/arch/sdl/dingoo-files/*.dge SDLVICE-dingoo-$VICEVERSION
 cp $TOPSRCDIR/src/arch/sdl/dingoo-files/sdl-vicerc* SDLVICE-dingoo-$VICEVERSION
 cp -a $TOPSRCDIR/data/C128 $TOPSRCDIR/data/C64 SDLVICE-dingoo-$VICEVERSION
