+++ vice/src/arch/sdl/make-bindist_gp2x.sh
@@ -11,20 +11,28 @@ VICEVERSION=$2
 ZIPKIND=$3
 TOPSRCDIR=$4
 
+if [ ! -e src/x64 -o ! -e src/x64dtv -o ! -e src/x64sc -o ! -e src/x128 -o ! -e src/xvic -o ! -e src/xpet -o ! -e src/xplus4 -o ! -e src/xcbm2 -o ! -e src/c1541 -o ! -e src/petcat -o ! -e src/cartconv ]
+then
+  echo Error: executable file\(s\) not found, do a \"make\" first
+  exit 1
+fi
+
 echo Generating GP2X SDL port binary distribution.
 rm -f -r SDLVICE-gp2x-$VICEVERSION
 mkdir SDLVICE-gp2x-$VICEVERSION
 $STRIP src/x64
 $STRIP src/x64dtv
+$STRIP src/x64sc
 $STRIP src/x128
 $STRIP src/xvic
 $STRIP src/xpet
 $STRIP src/xplus4
 $STRIP src/xcbm2
 cp src/x64 src/x64dtv src/x128 src/xvic SDLVICE-gp2x-$VICEVERSION
-cp src/xpet src/xplus4 src/xcbm2 SDLVICE-gp2x-$VICEVERSION
+cp src/xpet src/xplus4 src/xcbm2 src/x64sc SDLVICE-gp2x-$VICEVERSION
 cp $TOPSRCDIR/src/arch/unix/gp2x/gpe-files/x64.gpe SDLVICE-gp2x-$VICEVERSION
 cp $TOPSRCDIR/src/arch/unix/gp2x/gpe-files/x64dtv.gpe SDLVICE-gp2x-$VICEVERSION
+cp $TOPSRCDIR/src/arch/unix/gp2x/gpe-files/x64sc.gpe SDLVICE-gp2x-$VICEVERSION
 cp $TOPSRCDIR/src/arch/unix/gp2x/gpe-files/x128.gpe SDLVICE-gp2x-$VICEVERSION
 cp $TOPSRCDIR/src/arch/unix/gp2x/gpe-files/xvic.gpe SDLVICE-gp2x-$VICEVERSION
 cp $TOPSRCDIR/src/arch/unix/gp2x/gpe-files/xpet.gpe SDLVICE-gp2x-$VICEVERSION
